qcacld-3.0: Use hex dump instead of printing WOW pattern
Currently in the driver, we are printing the WOW pattern to the kernel
logs. This pattern is an array and printing it one by one can lead to
excessive logging.

Instead of printing one by one, use qdf_trace_hex_dump() to print the
pattern in one go. Also move the logging from kernel to the driver logs.

Change-Id: I6a3b02b0e57d8c27bb5ea4a7c44f67634c0f33fd
CRs-Fixed: 2661242