GKI: clk: qcom: debugcc: fix clk_measure register failure
For builtin drivers, the debugfs nodes are created in clk_debug_init()
by the late_initcall. For modules, the debugfs nodes will be created
when the clk_hw is registered during probe. The clk_measure debugfs node
in particular needs the clk_measure pointer registered before
registering the associated clk_hw so that the core clock driver can
setup the debugfs node for it during registration.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Bug: 160389301
Test: boot redfin with userdebug, ls /d/clk/*/clk_measure
Change-Id: I2ae8bc98b209d1269445d680ed9ae230a8893977
Signed-off-by: dreamisbaka <jolinux.g@gmail.com>